apps/listings-manager/requirements.txt
tf2-utilities
python-dotenv
uvicorn
aiohttp
fastapi
motor